+++ sql/README.md
@@ -2,6 +2,6 @@
 
 https://deno.land/x/database/sql
 
-**x/database/sql** provides an abstract interface for SQL databases in Deno,
-inspired by Go's [**database/sql**](https://pkg.go.dev/database/sql) and
-[**jmoiron/sqlx**](https://pkg.go.dev/github.com/jmoiron/sqlx) packages.
+**x/database/sql** provides an abstract TypeScript interface for SQL databases
+in Deno, inspired by Go's [**database/sql**](https://pkg.go.dev/database/sql)
+and [**jmoiron/sqlx**](https://pkg.go.dev/github.com/jmoiron/sqlx) packages.
