+++ tests/example.test.ts
@@ -1,0 +1,23 @@
+import "../sql/driver.ts";
+import * as sql from "../sql/sql.ts";
+import { SQL } from "../sql/strings.ts";
+import * as sqlite from "../x/sqlite.ts";
+
+Deno.test("example", async () => {
+  const db = sql.open(":memory:", sqlite);
+
+  const Users = SQL.identifier("Users");
+
+  const columnNames = ["firstName", "lastName"];
+  const columns = columnsNames.map((name) =>
+    SQL`${SQL.identifier(name)} varchar(32)`
+  );
+  await db.connect().use(async (connection) => {
+    await connection.exec(SQL`
+      CREATE TABLE IF NOT EXISTS
+      ${Users} (
+        ${SQL`, `.joining(columns)}
+      )
+    `);
+  });
+});
